added docker tls support and systemd timesync; fixed unicode error

+++ docker-tls.sh
@@ -10,32 +10,38 @@ fi
 cd "$(dirname $0)"
 mkdir -p docker-tls
 cd docker-tls
-if [[ ! -f ca.key ]]; then
-    openssl genrsa -out ca.key 4096
-fi
 if [[ ! -f ca.crt ]]; then
+    if [[ ! -f ca.key ]]; then
+        openssl genrsa -out ca.key 4096
+    fi
     openssl req -new -subj "/CN=ca" -x509 -days 9999 -key ca.key -out ca.crt
 fi
 
 create_yml() {
-    local host=$1
-    local ext=$2
+    local pillar=$1
+    local host=$2
+    local ext=$3
     local d="../srv/pillar/secrets/docker-tls"
+    local yml="$d/$host.yml"
+    if [[ -f $yml ]]; then
+        return
+    fi
     mkdir -p "$d"
     openssl genrsa -out "$host.key" 4096
     openssl req -subj "/CN=$host" -sha256 -new -key "$host.key" -out "$host.csr"
     echo "$ext" > "$host.cfg"
     openssl x509 -req -days 9999 -sha256 -in "$host.csr" -CA ca.crt -CAkey ca.key -CAcreateserial -out "$host.crt" -extfile "$host.cfg"
     rm -f "$host.csr" "$host.cfg"
-    python - <<EOF > "$d/$host.yml"
+    python - <<EOF > "$yml"
 from __future__ import print_function
 import re
 
-print('radia:\n  docker_service:')
+print('radia:\n  $pillar:')
 for k, f in ('tlskey', '$host.key'), ('tlscacert', 'ca.crt'), ('tlscert', '$host.crt'):
     print('    {}: |'.format(k))
     print(re.sub('^', '      ', open(f).read(), flags=re.MULTILINE))
 EOF
+    echo "Created: $yml"
 }
 
 for host in "$@"; do
@@ -44,9 +50,9 @@ for host in "$@"; do
         # Should only be used for testing
         host=server
     fi
-    create_yml "$host" "subjectAltName = IP:$ip,IP:127.0.0.1"
+    create_yml docker_service "$host" "subjectAltName = IP:$ip,IP:127.0.0.1"
 done
 
 if [[ ! -f client.key ]]; then
-    create_yml client 'extendedKeyUsage = clientAuth'
+    create_yml docker_master_client client 'extendedKeyUsage = clientAuth'
 fi
